Add files via upload
The data analysis reveals that the total claim amount is highest for private hospitals, followed by government hospitals and trust hospitals. The age of the patient, the type of procedure, and the hospital type are all significant predictors of the total claim amount.

The random forest regressor model outperforms the linear regression model in predicting the total claim amount. This suggests that the random forest regressor model is better able to capture the complex relationships between the different variables.

The data analysis and machine learning models can be used to help hospitals and insurance companies better predict the cost of care for different patients and procedures. This information can be used to set appropriate premiums and to develop strategies to reduce costs.